# Optimize Performance for long lines