--- a/composer.json
+++ b/composer.json
@@ -34,6 +34,9 @@
     "require": {
         "php": ">=5.3.0"
     },
+    "suggest": {
+        "ext-mbstring" : "Massive performance enhancement of line folding"
+    },
     "require-dev": {
         "phpunit/phpunit": "~4.3"
     },