--- a/src/Eluceo/iCal/Util/ComponentUtil.php
+++ b/src/Eluceo/iCal/Util/ComponentUtil.php
@@ -26,23 +26,17 @@ class ComponentUtil
      */
     public static function fold($string)
     {
-        $lines = array();
-        $array = preg_split('/(?<!^)(?!$)/u', $string);
-
-        $line   = '';
-        $lineNo = 0;
-        foreach ($array as $char) {
-            $charLen = strlen($char);
-            $lineLen = strlen($line);
-            if ($lineLen + $charLen > 75) {
-                $line = ' ' . $char;
-                ++$lineNo;
+        $lines = [];
+        while (strlen($string) > 0) {
+            if (strlen($string) > 75) {
+                $lines[] = mb_strcut($string, 0, 75, 'utf-8') . "\r\n";
+                $string = ' ' . mb_strcut($string, 75, strlen($string), 'utf-8');
             } else {
-                $line .= $char;
+                $lines[] = $string;
+                $string = '';
+                break;
             }
-            $lines[$lineNo] = $line;
         }
-
         return $lines;
     }
 }