--- a/src/Eluceo/iCal/Util/ComponentUtil.php
+++ b/src/Eluceo/iCal/Util/ComponentUtil.php
@@ -29,7 +29,7 @@ class ComponentUtil
         $lines = [];
         while (strlen($string) > 0) {
             if (strlen($string) > 75) {
-                $lines[] = mb_strcut($string, 0, 75, 'utf-8') . "\r\n";
+                $lines[] = mb_strcut($string, 0, 75, 'utf-8');
                 $string = ' ' . mb_strcut($string, 75, strlen($string), 'utf-8');
             } else {
                 $lines[] = $string;