use mb_strcut() instead of preg_split()