@stchr 2017-10-20
Is there anything I can do to push this PR?
It is an massive improvement both CPU and memory usage. Here a quick and dirty test with 1000 calls of 
```php
for ($i = 0; $i < 1000; $i++) {
    Eluceo\iCal\Util\ComponentUtil::fold('Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.');
}
```

Before:

Function|Call Count|Self Wall Time|Self CPU|Self MemoryUsage|Inclusive Wall Time|Inclusive CPU|Inclusive Memory Usage
--------|----------|--------------|--------|----------------|-------------------|-------------|----------------------
Eluceo\iCal\Util\ComponentUtil::fold|1000|1,863,972 µs|921,875 µs|-60,887,248 bytes|2,121,698 µs|2,125,000 µs|2,752 bytes
preg_split|1000|221,425 µs|921,875 µs|60,889,544 bytes|221,425 µs|250,000 µs|60,889,544 bytes
strlen|1182000|36,301 µs|921,875 µs|456 bytes|36,301 µs|953,125 µs|456 bytes

After:

Function|Call Count|Self Wall Time|Self CPU|Self MemoryUsage|Inclusive Wall Time|Inclusive CPU|Inclusive Memory Usage
--------|----------|--------------|--------|----------------|-------------------|-------------|----------------------
Eluceo\iCal\Util\ComponentUtil::fold|1000|57,037 µs|46,875 µs|-2,863,024 bytes|87,592 µs|93,750 µs|93,750 µs
mb_strcut|14000|29,701 µs|46,875 µs|2,864,456 bytes|29,701 µs|46,875 µs|2,864,456 bytes

It is nearly 25 times faster and uses 25 times less memory for this simple example. Even a fallback to the old implementation if multibyte extension is not installed is provided.