@markuspoerschke 2017-10-26
Hallo @stchr,

thanks for your contribution. Looks good. I tested the output and it was the same.

I also tested the performance using [PHPench](https://github.com/mre/PHPench). This is how it looks to me:

![image](https://user-images.githubusercontent.com/1222377/32049223-869982c2-ba4c-11e7-8f56-4efc88f23a40.png)

```
<?php

require_once __DIR__ . '/../vendor/autoload.php';

abstract class AbstractBenchmark implements \mre\PHPench\BenchmarkInterface
{
    protected $i = 1;
    protected $text = 'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.';

    function setUp($i)
    {
        $this->i = $i;
    }

    public function execute()
    {
        for ($i = 0; $i < $this->i; $i++) {
            $this->fold($this->text);
        }
    }

    abstract protected function fold($string);
}

class BenchmarkPregSplitStrLen extends AbstractBenchmark
{

    protected function fold($string)
    {
        $lines = array();
        $array = preg_split('/(?<!^)(?!$)/u', $string);
        $line = '';
        $lineNo = 0;
        foreach ($array as $char) {
            $charLen = strlen($char);
            $lineLen = strlen($line);
            if ($lineLen + $charLen > 75) {
                $line = ' ' . $char;
                ++$lineNo;
            } else {
                $line .= $char;
            }
            $lines[$lineNo] = $line;
        }
    }
}

class BenchmarkMbStrcut extends AbstractBenchmark
{
    protected function fold($string)
    {
        $lines = [];
        while (strlen($string) > 0) {
            if (strlen($string) > 75) {
                $lines[] = mb_strcut($string, 0, 75, 'utf-8');
                $string = ' ' . mb_strcut($string, 75, strlen($string), 'utf-8');
            } else {
                $lines[] = $string;
                $string = '';
                break;
            }
        }
    }
}

// Create a new benchmark instance
$phpench = new mre\PHPench(new \mre\PHPench\Aggregator\MedianAggregator);

$output = new \mre\PHPench\Output\GnuPlotOutput('ical_fold.png', 1024, 768);
$output->setTitle('Compare fold with preg_split and strln vs. mb_strcut');
$phpench->setOutput($output);


// Add your test to the instance
$phpench->addBenchmark(new BenchmarkPregSplitStrLen, 'preg_split + strlen');
$phpench->addBenchmark(new BenchmarkMbStrcut, 'mb_strcut');

// Run the benchmark and plot the results in realtime.
// With the second parameter you can specify
// the start, end and step for each call
$phpench->setInput(range(1, pow(2, 12), 512));
$phpench->setRepetitions(4);
$phpench->run(true);
```

It is really a very good performance improvement! Thanks for that.

I will merge the changes now. Thanks for your contribution! Very good catch! Looks like you are generating a lot of iCal files? 😄 

Viele Grüße
Markus