@markuspoerschke 2017-10-26
The change was released in `0.11.4`! 🍻