@stchr 2017-10-26
Thanks! Didn't know PHPench.
And yes, I'm generating large iCal subscriptions with a lot of VEVENT items 👏